+++ examples/cxx/LockExample.cpp
@@ -1,7 +1,7 @@
 /*-
  * See the file LICENSE for redistribution information.
  *
- * Copyright (c) 1997, 2014 Oracle and/or its affiliates.  All rights reserved.
+ * Copyright (c) 1997, 2016 Oracle and/or its affiliates.  All rights reserved.
  *
  * $Id$
  */
@@ -100,7 +100,7 @@ main(int argc, char *argv[])
 }
 
 LockExample::LockExample(const char *home, u_int32_t maxlocks, int do_unlink)
-:	DbEnv(0)
+:	DbEnv((u_int32_t)0)
 ,	ecode(0)
 {
 	int ret;
